refactor: types and constants out of utils
