+++ typescript/create-onchain-agent/src/types.ts
@@ -1,0 +1,23 @@
+
+export type EVMNetwork =
+  | "ethereum-mainnet"
+  | "ethereum-sepolia"
+  | "polygon-mainnet"
+  | "polygon-mumbai"
+  | "base-mainnet"
+  | "base-sepolia"
+  | "arbitrum-mainnet"
+  | "arbitrum-sepolia"
+  | "optimism-mainnet"
+  | "optimism-sepolia";
+
+export type SVMNetwork = "solana-mainnet" | "solana-devnet" | "solana-testnet";
+
+export type Network = EVMNetwork | SVMNetwork;
+
+export type WalletProviderChoice = 'CDP' | 'Viem' | 'Privy' | 'SolanaKeypair'
+
+export type WalletProviderRouteConfiguration = {
+    env: string[];
+    apiRoute: string
+}
